# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Smashgrade, a web application for tracking your student grades.",
        "title": "Smashgrade Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "HFTM Grenchen",
            "url": "https://www.hftm.ch"
        },
        "license": {
            "name": "Affero General Public License",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "1.0"
    },
    "host": "api.smashgrade.ch",
    "basePath": "/v1",
    "paths": {
...
        "/curriculumtypes": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all curriculum types",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "curriculumtypes"
                ],
                "summary": "Get all curriculum types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Curriculumtype"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    }
                }
            }
        },
        "/evaluationtypes": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all evaluation types",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evaluationtypes"
                ],
                "summary": "Get all evaluation types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Evaluationtype"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    }
                }
            }
        },
...
        "/examtypes": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all exam types",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "examtypes"
                ],
                "summary": "Get all exam types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Examtype"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    }
                }
            }
        },
        "/fields": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all fields",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fields"
                ],
                "summary": "Get all fields",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Field"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    }
                }
            }
        },
        "/focuses": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all focuses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "focuses"
                ],
                "summary": "Get all focuses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Focus"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    }
                }
            }
        },
        "/gradetypes": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all grade types",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gradetypes"
                ],
                "summary": "Get all grade types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Gradetype"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a gradetype",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gradetype"
                ],
                "summary": "Create a gradetype",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Gradetype"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Gradetype"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    }
                }
            }
        },
...
        "/states": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all states",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "states"
                ],
                "summary": "Get all states",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.State"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "error.ApiError": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.Course": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "exams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Exam"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "modules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Module"
                    }
                },
                "number": {
                    "type": "string"
                },
                "selectedCourses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SelectedCourse"
                    }
                },
                "teachedBy": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                },
                "updated": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "models.Curriculum": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "curriculumType": {
                    "$ref": "#/definitions/models.Curriculumtype"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "endValidity": {
                    "type": "string"
                },
                "focus": {
                    "$ref": "#/definitions/models.Focus"
                },
                "id": {
                    "type": "integer"
                },
                "modules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Module"
                    }
                },
                "startvalidity": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/models.State"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "models.Curriculumtype": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "durationYears": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "models.Evaluationtype": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "models.Exam": {
            "type": "object",
            "properties": {
                "course": {
                    "$ref": "#/definitions/models.Course"
                },
                "courseId": {
                    "type": "integer"
                },
                "courseVersion": {
                    "type": "integer"
                },
                "created": {
                    "type": "string"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "examtype": {
                    "$ref": "#/definitions/models.Examtype"
                },
                "id": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                },
                "weighting": {
                    "type": "integer"
                }
            }
        },
...
        "models.Examtype": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "models.Field": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.Focus": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "field": {
                    "$ref": "#/definitions/models.Field"
                },
                "id": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "models.Gradetype": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
...
        "models.Module": {
            "type": "object",
            "properties": {
                "courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Course"
                    }
                },
                "created": {
                    "type": "string"
                },
                "curriculums": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Curriculum"
                    }
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "evaluationType": {
                    "$ref": "#/definitions/models.Evaluationtype"
                },
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/models.State"
                },
                "studyStage": {
                    "$ref": "#/definitions/models.StudyStage"
                },
                "updated": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "models.Role": {
            "type": "object",
            "properties": {
                "claim": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.SelectedCourse": {
            "type": "object",
            "properties": {
                "classStartYear": {
                    "type": "string"
                },
                "courseId": {
                    "type": "integer"
                },
                "courseVersion": {
                    "type": "integer"
                },
                "dispensed": {
                    "type": "boolean"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "models.State": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moduleID": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
...
        "models.StudyStage": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moduleID": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "classStartYear": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "curriculumId": {
                    "type": "integer"
                },
                "deleted": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Field"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Role"
                    }
                },
                "selectedCourses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SelectedCourse"
                    }
                },
                "teachesCourses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Course"
                    }
                },
                "updated": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}